[FIX] point_of_sale: avoid loading uninstalled models
Models from uninstalled modules that were previously loaded
in the indexDB of a PoS config were raising a traceback
when opening the config again, as the PoS was still trying to
load them.

We now avoid trying to load them if they are not present anymore,
and drop the table from indexedDB.

task-6567097

+++ addons/point_of_sale/static/src/app/plugins/pos_data_plugin.js
@@ -159,6 +159,11 @@ export class PosDataPlugin extends Plugin {
             return [key, model];
         });
 
+        // Ensure previous connection is closed to avoid blocking silently
+        if (this.indexedDB?.db) {
+            this.indexedDB.db.close();
+        }
+
         return new Promise((resolve) => {
             this.indexedDB = new IndexedDB(this.databaseName, false, models, resolve, this.dialog);
         });
@@ -482,10 +487,15 @@ export class PosDataPlugin extends Plugin {
         // Remove data related to models previously loaded but not anymore.
         // This can happen when uninstalling a module.
         const allModelNames = Object.keys(data);
+        const existingStoreNames = new Set(
+            this.indexedDB.dbStores.map(([, storeName]) => storeName)
+        );
         for (const [model, values] of Object.entries(localData)) {
             if (!allModelNames.includes(model)) {
-                const idsToRemove = values.map((r) => r.id);
-                await this.indexedDB.delete(model, idsToRemove);
+                if (existingStoreNames.has(model)) {
+                    const idsToRemove = values.map((r) => r.id);
+                    await this.indexedDB.delete(model, idsToRemove);
+                }
                 delete localData[model];
             }
         }
